move ppcontentblock to content
