+++ assets/components/content/content.scss
