+++ assets/pages/digital-subscription-checkout/components/thankYouPendingContent.jsx
@@ -4,7 +4,7 @@
 
 import React from 'react';
 
-import ProductPageContentBlock from 'components/productPage/productPageContentBlock/productPageContentBlock';
+import Content from 'components/content/content';
 import Text, { LargeParagraph } from 'components/text/text';
 
 import { sendClickedEvent } from 'helpers/tracking/clickTracking';
@@ -16,7 +16,7 @@ function ThankYouPendingContent() {
 
   return (
     <div>
-      <ProductPageContentBlock>
+      <Content>
         <Text>
           <LargeParagraph>
             Thank you for subscribing to the Digital Pack.
@@ -42,7 +42,7 @@ function ThankYouPendingContent() {
             )} for customer support.
           </p>
         </Text>
-      </ProductPageContentBlock>
+      </Content>
     </div>
   );
